+++ src/TRegx/SafeRegex/Guard/Arrays.php
@@ -9,4 +9,9 @@ class Arrays
     {
         return !array_diff_key($array1, $array2) && !array_diff_key($array2, $array1);
     }
+
+    public static function flatten(array $array): array
+    {
+        return call_user_func_array('array_merge', $array);
+    }
 }